fix: add Windows daemon support
- Add platform-specific daemon launcher using CREATE_NEW_PROCESS_GROUP to
  isolate the daemon from parent console Ctrl+C events
- Split syncDir into unix/windows builds: no-op on Windows since
  FlushFileBuffers requires write access on directory handles
- Introduce lockPIDPath helper with a .pid sidecar on Windows to avoid
  LockFileEx byte-range lock conflict when writing the PID
- Fix SQLite DSN for Windows drive letters (ensure triple-slash URI form)
