+++ internal/cli/daemon_launch_other.go
@@ -1,4 +1,4 @@
-//go:build !(darwin || dragonfly || freebsd || linux || netbsd || openbsd)
+//go:build !(darwin || dragonfly || freebsd || linux || netbsd || openbsd || windows)
 
 package cli
